Move specimen permissions into specimen module
