+++ specimen/src/org/labkey/specimen/SpecimenModule.java
@@ -42,6 +42,7 @@ import org.labkey.api.query.QueryUpdateService;
 import org.labkey.api.query.QueryView;
 import org.labkey.api.query.UserSchema;
 import org.labkey.api.security.User;
+import org.labkey.api.security.permissions.Permission;
 import org.labkey.api.security.roles.RoleManager;
 import org.labkey.api.specimen.SpecimenMigrationService;
 import org.labkey.api.specimen.SpecimenQuerySchema;
@@ -52,10 +53,16 @@ import org.labkey.api.study.SpecimenService;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyInternalService;
 import org.labkey.api.study.StudyService;
+import org.labkey.api.study.StudyUrls;
 import org.labkey.api.study.importer.SimpleStudyImportContext;
 import org.labkey.api.study.importer.SimpleStudyImporterRegistry;
 import org.labkey.api.study.writer.SimpleStudyWriterRegistry;
 import org.labkey.api.usageMetrics.UsageMetricsService;
+import org.labkey.api.util.HtmlString;
+import org.labkey.api.util.HtmlStringBuilder;
+import org.labkey.api.util.LinkBuilder;
+import org.labkey.api.util.PageFlowUtil;
+import org.labkey.api.util.SafeToRender;
 import org.labkey.api.util.SystemMaintenance;
 import org.labkey.api.util.emailTemplate.EmailTemplateService;
 import org.labkey.api.util.logging.LogHelper;
@@ -74,12 +81,14 @@ import org.labkey.specimen.importer.SpecimenSchemaImporter;
 import org.labkey.specimen.importer.SpecimenSettingsImporter;
 import org.labkey.specimen.model.SpecimenRequestEventType;
 import org.labkey.specimen.pipeline.SpecimenPipeline;
-import org.labkey.specimen.requirements.SpecimenRequestRequirementProvider;
 import org.labkey.specimen.query.SpecimenPivotByDerivativeType;
 import org.labkey.specimen.query.SpecimenPivotByPrimaryType;
 import org.labkey.specimen.query.SpecimenPivotByRequestingLocation;
 import org.labkey.specimen.query.SpecimenQueryView;
 import org.labkey.specimen.query.SpecimenUpdateService;
+import org.labkey.specimen.requirements.SpecimenRequestRequirementProvider;
+import org.labkey.specimen.security.permissions.EditSpecimenDataPermission;
+import org.labkey.specimen.security.permissions.ManageRequestSettingsPermission;
 import org.labkey.specimen.security.roles.SpecimenCoordinatorRole;
 import org.labkey.specimen.security.roles.SpecimenRequesterRole;
 import org.labkey.specimen.settings.RepositorySettings;
@@ -233,6 +242,30 @@ public class SpecimenModule extends SpringModule
                     return new SpecimenPivotByRequestingLocation(schema, study, cf);
                 }
                 return null;
+            }
+
+            @Override
+            public @NotNull SafeToRender getStudySummaryLink(Container c, User user)
+            {
+                // Must have ManageRequestSettingsPermission and requests enabled (otherwise non-admin will see a blank Manage Study page)
+                if (c.hasPermission(user, ManageRequestSettingsPermission.class) && isEnableRequests(c))
+                {
+                    return HtmlStringBuilder.of()
+                        .unsafeAppend("<p>")
+                        .append(LinkBuilder.labkeyLink(
+                            "Manage Specimen Request Settings",
+                            PageFlowUtil.urlProvider(StudyUrls.class).getManageStudyURL(c)
+                        ))
+                        .unsafeAppend("</p>");
+                }
+
+                return HtmlString.EMPTY_STRING;
+            }
+
+            @Override
+            public @NotNull Class<? extends Permission> getSpecimenEditDataPermission()
+            {
+                return EditSpecimenDataPermission.class;
             }
         });
