+++ specimen/src/org/labkey/specimen/actions/SpecimenApiController.java
@@ -33,8 +33,6 @@ import org.labkey.api.security.permissions.ReadPermission;
 import org.labkey.api.specimen.Vial;
 import org.labkey.api.specimen.location.LocationImpl;
 import org.labkey.api.specimen.location.LocationManager;
-import org.labkey.specimen.model.SpecimenTypeSummary;
-import org.labkey.api.specimen.security.permissions.RequestSpecimensPermission;
 import org.labkey.api.study.StudyUtils;
 import org.labkey.api.util.PageFlowUtil;
 import org.labkey.api.view.NotFoundException;
@@ -48,8 +46,10 @@ import org.labkey.specimen.importer.RequestabilityManager;
 import org.labkey.specimen.model.AdditiveType;
 import org.labkey.specimen.model.DerivativeType;
 import org.labkey.specimen.model.PrimaryType;
+import org.labkey.specimen.model.SpecimenTypeSummary;
 import org.labkey.specimen.requirements.SpecimenRequest;
 import org.labkey.specimen.security.permissions.ManageRequestsPermission;
+import org.labkey.specimen.security.permissions.RequestSpecimensPermission;
 import org.labkey.specimen.settings.RepositorySettings;
 import org.labkey.specimen.settings.SettingsManager;
 import org.springframework.validation.BindException;
