+++ specimen/src/org/labkey/specimen/actions/SpecimenController.java
@@ -107,7 +107,6 @@ import org.labkey.api.security.permissions.AbstractContainerScopingTest;
 import org.labkey.api.security.permissions.AdminPermission;
 import org.labkey.api.security.permissions.ReadPermission;
 import org.labkey.api.security.permissions.UpdatePermission;
-import org.labkey.api.security.roles.ReaderRole;
 import org.labkey.api.specimen.SpecimenQuerySchema;
 import org.labkey.api.specimen.SpecimenSchema;
 import org.labkey.api.specimen.Vial;
@@ -115,9 +114,6 @@ import org.labkey.api.specimen.location.LocationImpl;
 import org.labkey.api.specimen.location.LocationManager;
 import org.labkey.api.specimen.model.SpecimenComment;
 import org.labkey.api.specimen.model.SpecimenTablesProvider;
-import org.labkey.api.specimen.security.permissions.EditSpecimenDataPermission;
-import org.labkey.api.specimen.security.permissions.ManageRequestSettingsPermission;
-import org.labkey.api.specimen.security.permissions.RequestSpecimensPermission;
 import org.labkey.api.study.CohortFilter;
 import org.labkey.api.study.Dataset;
 import org.labkey.api.study.MapArrayExcelWriter;
@@ -171,11 +167,10 @@ import org.labkey.specimen.AmbiguousLocationException;
 import org.labkey.specimen.RequestEventType;
 import org.labkey.specimen.RequestedSpecimens;
 import org.labkey.specimen.SpecimenManager;
-import org.labkey.specimen.SpecimenRequestException;
 import org.labkey.specimen.SpecimenModule;
+import org.labkey.specimen.SpecimenRequestException;
 import org.labkey.specimen.SpecimenRequestManager;
 import org.labkey.specimen.SpecimenRequestStatus;
-import org.labkey.specimen.security.roles.SpecimenRequesterRole;
 import org.labkey.specimen.importer.QueryBasedSpecimenTransform;
 import org.labkey.specimen.importer.RequestabilityManager;
 import org.labkey.specimen.importer.SimpleSpecimenImporter;
@@ -194,14 +189,18 @@ import org.labkey.specimen.requirements.SpecimenRequest;
 import org.labkey.specimen.requirements.SpecimenRequestRequirement;
 import org.labkey.specimen.requirements.SpecimenRequestRequirementProvider;
 import org.labkey.specimen.requirements.SpecimenRequestRequirementType;
+import org.labkey.specimen.security.permissions.EditSpecimenDataPermission;
 import org.labkey.specimen.security.permissions.ManageDisplaySettingsPermission;
 import org.labkey.specimen.security.permissions.ManageNewRequestFormPermission;
 import org.labkey.specimen.security.permissions.ManageNotificationsPermission;
 import org.labkey.specimen.security.permissions.ManageRequestRequirementsPermission;
+import org.labkey.specimen.security.permissions.ManageRequestSettingsPermission;
 import org.labkey.specimen.security.permissions.ManageRequestStatusesPermission;
 import org.labkey.specimen.security.permissions.ManageRequestsPermission;
 import org.labkey.specimen.security.permissions.ManageSpecimenActorsPermission;
+import org.labkey.specimen.security.permissions.RequestSpecimensPermission;
 import org.labkey.specimen.security.permissions.SetSpecimenCommentsPermission;
+import org.labkey.specimen.security.roles.SpecimenRequesterRole;
 import org.labkey.specimen.settings.DisplaySettings;
 import org.labkey.specimen.settings.RepositorySettings;
 import org.labkey.specimen.settings.RequestNotificationSettings;
