+++ specimen/src/org/labkey/specimen/query/SpecimenQueryView.java
@@ -50,8 +50,6 @@ import org.labkey.api.specimen.SpecimenQuerySchema;
 import org.labkey.api.specimen.SpecimenSchema;
 import org.labkey.api.specimen.Vial;
 import org.labkey.api.specimen.model.SpecimenTablesProvider;
-import org.labkey.api.specimen.security.permissions.EditSpecimenDataPermission;
-import org.labkey.api.specimen.security.permissions.RequestSpecimensPermission;
 import org.labkey.api.study.CohortFilter;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyService;
@@ -66,6 +64,8 @@ import org.labkey.api.view.DataView;
 import org.labkey.api.view.ViewContext;
 import org.labkey.api.writer.HtmlWriter;
 import org.labkey.specimen.actions.SpecimenController;
+import org.labkey.specimen.security.permissions.EditSpecimenDataPermission;
+import org.labkey.specimen.security.permissions.RequestSpecimensPermission;
 import org.labkey.specimen.settings.DisplaySettings;
 import org.labkey.specimen.settings.RepositorySettings;
 import org.labkey.specimen.settings.SettingsManager;
