+++ specimen/src/org/labkey/specimen/security/permissions/EditSpecimenDataPermission.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2009-2026 LabKey Corporation
+ * Copyright (c) 2013-2026 LabKey Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,22 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.labkey.api.specimen.security.permissions;
+package org.labkey.specimen.security.permissions;
 
 import org.labkey.api.security.permissions.AbstractPermission;
-import org.labkey.api.study.StudyService;
+import org.labkey.specimen.SpecimenModule;
 
-/*
-* User: Dave
-* Date: May 18, 2009
-* Time: 12:01:15 PM
-*/
-public class ManageRequestSettingsPermission extends AbstractPermission
+public class EditSpecimenDataPermission extends AbstractPermission
 {
-    public ManageRequestSettingsPermission()
+    public EditSpecimenDataPermission()
     {
-        super("Manage Specimen Request and Tracking Settings",
-                "Allows management of specimen request and tracking settings",
-                StudyService.get().getStudyModuleClass()); // TODO: Switch to SpecimenModule.class
+        super(
+            "Edit Specimen Data",
+            "Allows editing of specimen data",
+            SpecimenModule.class
+        );
     }
 }
