+++ specimen/src/org/labkey/specimen/security/roles/SpecimenCoordinatorRole.java
@@ -17,18 +17,18 @@ package org.labkey.specimen.security.roles;
 
 import org.jetbrains.annotations.NotNull;
 import org.labkey.api.security.permissions.EditSharedViewPermission;
-import org.labkey.api.specimen.security.permissions.EditSpecimenDataPermission;
-import org.labkey.api.specimen.security.permissions.ManageRequestSettingsPermission;
-import org.labkey.api.specimen.security.permissions.RequestSpecimensPermission;
 import org.labkey.api.study.security.permissions.ManageStudyPermission;
+import org.labkey.specimen.security.permissions.EditSpecimenDataPermission;
 import org.labkey.specimen.security.permissions.LockSpecimensPermission;
 import org.labkey.specimen.security.permissions.ManageDisplaySettingsPermission;
 import org.labkey.specimen.security.permissions.ManageNewRequestFormPermission;
 import org.labkey.specimen.security.permissions.ManageNotificationsPermission;
 import org.labkey.specimen.security.permissions.ManageRequestRequirementsPermission;
+import org.labkey.specimen.security.permissions.ManageRequestSettingsPermission;
 import org.labkey.specimen.security.permissions.ManageRequestStatusesPermission;
 import org.labkey.specimen.security.permissions.ManageRequestsPermission;
 import org.labkey.specimen.security.permissions.ManageSpecimenActorsPermission;
+import org.labkey.specimen.security.permissions.RequestSpecimensPermission;
 import org.labkey.specimen.security.permissions.SetSpecimenCommentsPermission;
 
 import java.util.Collection;
