+++ specimen/src/org/labkey/specimen/view/SpecimenToolsWebPartFactory.java
@@ -15,7 +15,6 @@
  */
 package org.labkey.specimen.view;
 
-import org.labkey.api.specimen.security.permissions.RequestSpecimensPermission;
 import org.labkey.api.study.StudyUrls;
 import org.labkey.api.study.security.permissions.ManageStudyPermission;
 import org.labkey.api.study.view.StudyToolsWebPart;
@@ -27,6 +26,7 @@ import org.labkey.api.view.WebPartFactory;
 import org.labkey.specimen.actions.ShowSearchAction;
 import org.labkey.specimen.actions.SpecimenController.AutoReportListAction;
 import org.labkey.specimen.actions.SpecimenController.ShowCreateSpecimenRequestAction;
+import org.labkey.specimen.security.permissions.RequestSpecimensPermission;
 import org.labkey.specimen.settings.SettingsManager;
 
 import java.util.ArrayList;
