+++ specimen/src/org/labkey/specimen/view/manageSpecimens.jsp
@@ -15,13 +15,12 @@
  * limitations under the License.
  */
 %>
-<%@ page import="com.google.common.collect.Iterables"%>
+<%@ page import="com.google.common.collect.Iterables" %>
 <%@ page import="org.labkey.api.data.Container" %>
 <%@ page import="org.labkey.api.exp.property.Domain" %>
 <%@ page import="org.labkey.api.security.User" %>
 <%@ page import="org.labkey.api.security.permissions.AdminPermission" %>
 <%@ page import="org.labkey.api.specimen.model.SpecimenTablesProvider" %>
-<%@ page import="org.labkey.api.specimen.security.permissions.ManageRequestSettingsPermission" %>
 <%@ page import="org.labkey.api.study.SpecimenService" %>
 <%@ page import="org.labkey.api.study.SpecimenTransform" %>
 <%@ page import="org.labkey.api.study.Study" %>
@@ -41,6 +40,7 @@
 <%@ page import="org.labkey.specimen.actions.SpecimenController.ManageSpecimenWebPartAction" %>
 <%@ page import="org.labkey.specimen.actions.SpecimenController.ManageStatusesAction" %>
 <%@ page import="org.labkey.specimen.requirements.SpecimenRequestRequirementProvider" %>
+<%@ page import="org.labkey.specimen.security.permissions.ManageRequestSettingsPermission" %>
 <%@ page import="org.labkey.specimen.settings.SettingsManager" %>
 <%@ page import="java.util.Collection" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
