+++ specimen/src/org/labkey/specimen/view/webPart.jsp
@@ -15,10 +15,9 @@
  * limitations under the License.
  */
 %>
-<%@ page import="org.labkey.api.data.Container"%>
-<%@ page import="org.labkey.api.security.User"%>
-<%@ page import="org.labkey.api.security.permissions.AdminPermission"%>
-<%@ page import="org.labkey.api.specimen.security.permissions.RequestSpecimensPermission"%>
+<%@ page import="org.labkey.api.data.Container" %>
+<%@ page import="org.labkey.api.security.User" %>
+<%@ page import="org.labkey.api.security.permissions.AdminPermission" %>
 <%@ page import="org.labkey.api.study.StudyUrls" %>
 <%@ page import="org.labkey.api.util.HtmlString" %>
 <%@ page import="org.labkey.api.view.ActionURL" %>
@@ -29,6 +28,7 @@
 <%@ page import="org.labkey.specimen.actions.SpecimenController.AutoReportListAction" %>
 <%@ page import="org.labkey.specimen.actions.SpecimenController.ShowCreateSpecimenRequestAction" %>
 <%@ page import="org.labkey.specimen.actions.SpecimenController.ViewRequestsAction" %>
+<%@ page import="org.labkey.specimen.security.permissions.RequestSpecimensPermission" %>
 <%@ page import="org.labkey.specimen.settings.SettingsManager" %>
 <%@ page import="org.labkey.specimen.view.SpecimenWebPart" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
